+++ lib/vagrant-mutagen-utilize/action/cache_config.rb
