+++ lib/vagrant-mutagen-utilize/action/remove_config.rb
