+++ lib/vagrant-mutagen-utilize/plugin.rb
@@ -1,8 +1,8 @@
-require 'vagrant-mutagen-utilize/Action/UpdateConfig'
-require 'vagrant-mutagen-utilize/Action/CacheConfig'
-require 'vagrant-mutagen-utilize/Action/RemoveConfig'
-require 'vagrant-mutagen-utilize/Action/StartOrchestration'
-require 'vagrant-mutagen-utilize/Action/TerminateOrchestration'
+require 'vagrant-mutagen-utilize/action/update_config'
+require 'vagrant-mutagen-utilize/action/cache_config'
+require 'vagrant-mutagen-utilize/action/remove_config'
+require 'vagrant-mutagen-utilize/action/start_orchestration'
+require 'vagrant-mutagen-utilize/action/terminate_orchestration'
 
 module VagrantPlugins
   module Mutagen
